+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Deleting story used to cause the app to crash, that should be ended!
  - Disabling viewing story from profile when a user is in a room.
  - Preventing an issue that caused certain private rooms to have groups associated to them.
+ - Bugs fixed revolving around story creation.
 
 ## [1.6.0] - 2020-12-15